requirements.txt
azure-identity
openai
python-dotenv
langchain
langchain-openai
llama-index-llms-azure-openai
llama-index-llms-openai
llama-index-llms-openai-like
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-readers-file